feat: use mjs so env vars can be validated on build time
