+++ src/browserEnv.mjs
@@ -1,22 +1,22 @@
 import { z } from 'zod'
 
-// Coerces a string to true if it's "true" or "1", false if "false" or "0"
+// Coerces a string to true if it's "true", false if "false".
 export const coerceBoolean = z
-  .enum(['0', '1', 'true', 'false', 'TRUE', 'FALSE'])
-  .catch('false')
-  .transform((value) => value.toLowerCase() === 'true' || value === '1')
+  .string()
+  // only allow "true" or "false"
+  .refine((s) => s === 'true' || s === 'false')
+  // transform to boolean
+  .transform((s) => s === 'true')
+  // make sure tranform worked
+  .pipe(z.boolean())
 
 export const browserEnvSchema = z.object({
   NEXT_PUBLIC_ENABLE_STORAGE: coerceBoolean.default('false'),
   NEXT_PUBLIC_APP_NAME: z.string().default('Starter Kit'),
 })
 
-type BrowserEnv = {
-  [k in keyof z.infer<typeof browserEnvSchema>]: string | undefined
-}
-
 // Must add public env variables here explicitly so NextJS knows to expose them.
-const _browserEnv: BrowserEnv = {
+const _browserEnv = {
   NEXT_PUBLIC_ENABLE_STORAGE: process.env.NEXT_PUBLIC_ENABLE_STORAGE,
   NEXT_PUBLIC_APP_NAME: process.env.NEXT_PUBLIC_APP_NAME,
 }
@@ -28,7 +28,6 @@ if (!parsedEnv.success) {
     '❌ Invalid environment variables:',
     JSON.stringify(parsedEnv.error.format(), null, 4)
   )
-  process.exit(1)
 }
 
 /**
