+++ src/lib/r2.ts
@@ -4,7 +4,7 @@ import {
   type PutObjectCommandInput,
 } from '@aws-sdk/client-s3'
 import { getSignedUrl } from '@aws-sdk/s3-request-presigner'
-import { env } from '~/server/env'
+import { env } from '~/env.mjs'
 
 const { R2_ACCESS_KEY_ID, R2_SECRET_ACCESS_KEY, R2_ACCOUNT_ID } = env
