+++ src/pages/sign-in.tsx
@@ -3,7 +3,7 @@ import { Box, Flex, Skeleton, Text } from '@chakra-ui/react'
 import { RestrictedGovtMasthead } from '@opengovsg/design-system-react'
 import { noop } from 'lodash'
 import { useRouter } from 'next/router'
-import { browserEnv } from '~/browserEnv'
+import { browserEnv } from '~/browserEnv.mjs'
 import { MiniFooter } from '~/components/Footer/MiniFooter'
 import Suspense from '~/components/Suspense/Suspense'
 import { CALLBACK_URL_KEY } from '~/constants/params'
