+++ src/server/prisma.ts
@@ -2,8 +2,8 @@
  * Instantiates a single instance PrismaClient and save it on the global object.
  * @link https://www.prisma.io/docs/support/help-articles/nextjs-prisma-client-dev-practices
  */
-import { env } from './env'
 import { PrismaClient } from '@prisma/client'
+import { env } from '~/env.mjs'
 
 const prismaGlobal = global as typeof global & {
   prisma?: PrismaClient
